Fix dropdown and tabs location
