+++ src/components/Dropdown.tsx
@@ -1,5 +1,5 @@
-import React, { useState } from "react";
-import { ChevronDown } from "lucide-react";
+import React, { useState } from 'react';
+import { ChevronDown } from 'lucide-react';
 
 interface DropdownProps {
   label: string;
@@ -12,20 +12,20 @@ const Dropdown: React.FC<DropdownProps> = ({ label, items }) => {
 
   return (
     <div className="relative w-64">
-      <label className="block text-sm mb-1">Label</label>
+      <label className="mb-1 block text-sm">Label</label>
       <button
-        className="flex justify-between items-center w-full px-4 py-2 border border-gray-400 rounded-md bg-white focus:outline-none focus:ring-2 focus:ring-purple-500"
+        className="flex w-full items-center justify-between rounded-md border border-gray-400 bg-white px-4 py-2 focus:outline-none focus:ring-2 focus:ring-purple-500"
         onClick={() => setIsOpen(!isOpen)}
       >
         {selected}
         <ChevronDown size={18} />
       </button>
       {isOpen && (
-        <ul className="absolute left-0 mt-1 w-full bg-white border border-gray-300 rounded-md shadow-lg">
+        <ul className="absolute left-0 mt-1 w-full rounded-md border border-gray-300 bg-white shadow-lg">
           {items.map((item, index) => (
             <li
               key={index}
-              className="px-4 py-2 cursor-pointer hover:bg-gray-100"
+              className="cursor-pointer px-4 py-2 hover:bg-gray-100"
               onClick={() => {
                 setSelected(item);
                 setIsOpen(false);
